Extract filter program logic into dedicated module
